requirements.txt
pandas
SQLAlchemy
mysql-connector-python
openpyxl
python-dotenv